extracted calendar and donate route components
